refactor: extract parser state types into type/Parser2State.hpp
